+++ relecov_tools/build_schema.py
@@ -255,6 +255,8 @@ class BuildSchema(BaseModule):
     def validate_database_definition(self, json_data):
         """Validate the mandatory features and ensure:
         - No duplicate enum values in the JSON schema.
+        - All fields have an example.
+        - All examples should have the same type as JSON schema.
         - Date formats follow 'YYYY-MM-DD'.
 
         Args:
@@ -263,11 +265,14 @@ class BuildSchema(BaseModule):
         Returns:
             dict: A dictionary containing errors found, categorized by:
                 - Missing features
+                - Missing examples
                 - Duplicate enums
+                - Invalid example types
                 - Incorrect date formats
         """
         log_errors = {
             "missing_features": {},
+            "missing_examples": {},
             "duplicate_enums": {},
             "invalid_example_types": {},
             "invalid_date_formats": {},
@@ -309,24 +314,36 @@ class BuildSchema(BaseModule):
                         ]
                         log_errors["duplicate_enums"][prop_name] = duplicates
 
-            # Check date format for properties with type=string and format=date
-            if (
-                prop_features["type"] == "string"
-                and prop_features.get("format") == "date"
-            ):
-                example = prop_features.get("examples")
-                if example:
-                    if isinstance(example, datetime):
-                        example = example.strftime("%Y-%m-%d")
-                    if isinstance(example, str):
-                        try:
-                            datetime.strptime(example, "%Y-%m-%d")
-                        except ValueError:
-                            if prop_name not in log_errors["invalid_date_formats"]:
-                                log_errors["invalid_date_formats"][prop_name] = []
-                            log_errors["invalid_date_formats"][prop_name].append(
-                                f"Invalid date format '{example}', expected 'YYYY-MM-DD'"
-                            )
+            # Check for missing examples
+            example = prop_features.get("examples")
+            if example is None:
+                log_errors["missing_examples"][prop_name] = ["Missing example."]
+
+            feature_type = prop_features["type"]
+            match feature_type:
+                # Check date format for properties with type=string and format=date
+                case "string":
+                    if prop_features.get("format") == "date":
+                        if isinstance(example, datetime):
+                            example = example.strftime("%Y-%m-%d")
+                        if isinstance(example, str):
+                            try:
+                                datetime.strptime(example, "%Y-%m-%d")
+                            except ValueError:
+                                if prop_name not in log_errors["invalid_date_formats"]:
+                                    log_errors["invalid_date_formats"][prop_name] = []
+                                log_errors["invalid_date_formats"][prop_name].append(
+                                    f"Invalid date format '{example}', expected 'YYYY-MM-DD'"
+                                )
+
+                case "integer" | "number":
+                    function_to_convert = float if feature_type == "number" else int
+                    try:
+                        example = function_to_convert(example)
+                    except ValueError:
+                        log_errors["invalid_example_types"][prop_name] = [
+                            f"Value {example} is not a valid {feature_type}"
+                        ]
 
         # return log errors if any
         if any(log_errors.values()):
@@ -463,9 +480,14 @@ class BuildSchema(BaseModule):
                 items = value.split("; ")
                 if remove_ontology and target_key == "enum":
                     items = [re.sub(r"\s*\[.*?\]", "", item).strip() for item in items]
-                json_dict[target_key] = items if target_key == "enum" else [value]
+                json_dict[target_key] = items
         elif target_key == "description":
             json_dict[target_key] = handle_nan(data_dict.get(target_key, ""))
+            json_dict[target_key] = json_dict[target_key].strip()
+            if not json_dict[target_key].endswith(".") and not json_dict[
+                "description"
+            ].endswith(")"):
+                json_dict[target_key] = f"{json_dict[target_key]}."
         else:
             json_dict[target_key] = handle_nan(data_dict.get(target_key, ""))
         return json_dict
@@ -561,6 +583,9 @@ class BuildSchema(BaseModule):
 
             common_lab_enum = "; ".join(self._lab_uniques["collecting_institution"])
 
+            # Define definitions property. This will hold all $refs for the properties.
+            definitions = {"enums": {}}
+
             # Read property_ids in the database.
             #   Perform checks and create (for each property) feature object like:
             #       {'example':'A', 'ontology': 'B'...}.
@@ -574,7 +599,7 @@ class BuildSchema(BaseModule):
                     "submitting_institution",
                     "sequencing_institution",
                 ):
-                    db_features_dic["enum"] = common_lab_enum
+                    definitions["enums"][property_id] = {"enum": common_lab_enum}
 
                 # Parse property_ids that needs to be incorporated as complex fields in json_schema
                 if json_data[property_id].get("complex_field (Y/N)") == "Y":
@@ -626,6 +651,33 @@ class BuildSchema(BaseModule):
                                             pass
                                         options_dict[key] = value
                                 schema_property.update(options_dict)
+                        elif db_feature_key == "enum":
+                            enums_value = self.standard_jsonschema_object(
+                                db_features_dic, db_feature_key, remove_ontology=False
+                            )
+                            if enums_value:
+                                definitions["enums"][property_id] = enums_value
+                                reference = {"$ref": f"#/$defs/enums/{property_id}"}
+                                schema_property.update(reference)
+                        elif db_feature_key == "examples":
+                            examples_value = self.standard_jsonschema_object(
+                                db_features_dic, db_feature_key, remove_ontology=False
+                            )
+                            if db_features_dic["type"] != "string":
+                                # Examples for integer/number fields should be consistent.
+                                try:
+                                    examples_value[db_feature_key] = [
+                                        float(x) for x in examples_value[db_feature_key]
+                                    ]
+                                    examples_value[db_feature_key] = [
+                                        int(x) if x.is_integer() else x
+                                        for x in examples_value[db_feature_key]
+                                    ]
+                                except ValueError:
+                                    pass
+                            schema_property[schema_feature_key] = examples_value[
+                                db_feature_key
+                            ]
                         else:
                             std_json_feature = self.standard_jsonschema_object(
                                 db_features_dic, db_feature_key, remove_ontology=False
@@ -645,6 +697,7 @@ class BuildSchema(BaseModule):
                         required_property_unique.append(key)
             # TODO: So far it appears at the end of the new json schema. Ideally it should be placed before the properties statement.
             new_schema["required"] = required_property_unique
+            new_schema["$defs"] = definitions
             grouped_anyof = {}
 
             for prop_id, prop_data in json_data.items():
@@ -861,6 +914,7 @@ class BuildSchema(BaseModule):
             ]
             required_properties = set(json_schema.get("required", []))
             schema_properties = json_schema.get("properties")
+            enum_defs = json_schema.get("$defs", {}).get("enums", {})
 
             try:
                 schema_properties_flatten = relecov_tools.assets.schema_utils.metadatalab_template.schema_to_flatten_json(
@@ -890,13 +944,21 @@ class BuildSchema(BaseModule):
                 def clean_ontologies(enums):
                     return [re.sub(r"\s*\[.*?\]", "", item).strip() for item in enums]
 
-                df["enum"] = df["enum"].apply(
-                    lambda enum_list: (
-                        clean_ontologies(enum_list)
-                        if isinstance(enum_list, list)
-                        else enum_list
+                def resolve_enum_ref(ref: str, enum_defs: dict) -> list[str]:
+                    property_id = ref.split("/")[-1]
+                    values = enum_defs[property_id]["enum"]
+                    return (
+                        clean_ontologies(values) if isinstance(values, list) else values
+                    )
+
+                df["enum"] = df["$ref"].apply(
+                    lambda row: (
+                        resolve_enum_ref(row, enum_defs=enum_defs)
+                        if not pd.isna(row)
+                        else row
                     )
                 )
+
                 common_dropdown = self._lab_dropdowns["collecting_institution"]
 
                 lab_fields = [